requirements.txt
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.3.0
python-dotenv
termcolor
colorama
requests
beautifulsoup4
ddgs
markitdown
PyYAML
pydantic>=2.0.0
langfuse>=2.0.0
textual>=0.40.0
thefuzz>=0.19.0
rich>=13.0.0
browser-use>=0.1.0
playwright>=1.40.0
nest-asyncio>=1.5.0
python-socks>=2.0.0
